Move performance profiles data into separate class, update mode in deploy.sh
