+++ src/main/java/com/autotune/analyzer/Analyzer.java
@@ -21,6 +21,7 @@ import com.autotune.analyzer.exceptions.K8sTypeNotSupportedException;
 import com.autotune.analyzer.exceptions.MonitoringAgentNotFoundException;
 import com.autotune.analyzer.exceptions.MonitoringAgentNotSupportedException;
 import com.autotune.analyzer.services.*;
+import com.autotune.common.performanceProfiles.PerformanceProfilesDeployment;
 import com.autotune.utils.ServerContext;
 import org.eclipse.jetty.servlet.ServletContextHandler;
 
@@ -40,7 +41,7 @@ public class Analyzer {
         try {
             addServlets(contextHandler);
             AutotuneDeployment.getAutotuneObjects(autotuneDeployment);
-            AutotuneDeployment.getPerformanceProfiles();
+            PerformanceProfilesDeployment.getPerformanceProfiles();
         } catch (Exception e) {
             e.printStackTrace();
         }
